gui classes continued
