+++ note-application.iml
@@ -5,7 +5,4 @@
       <sourceFolder url="file://$MODULE_DIR$/src/main" isTestSource="false" />
     </content>
   </component>
-  <component name="CheckStyle-IDEA-Module" serialisationVersion="2">
-    <option name="activeLocationsIds" />
-  </component>
 </module>
