+++ src/main/java/view/BookmarksView.java
@@ -1,9 +1,11 @@
 package view;
 
+import app.BuildAppWindow;
+
 import javax.swing.*;
 
-import static view.BuildAppWindow.buttonClickable;
-import static view.BuildAppWindow.buttonUnclickable;
+import static app.BuildAppWindow.buttonClickable;
+import static app.BuildAppWindow.buttonUnclickable;
 
 /**
  * The LoginView for when the user has just started the program, and not logged in.
