--- src/main/java/view/LogoutView.java
@@ -1,17 +1,0 @@
-package view;
-
-import javax.swing.*;
-
-import static view.BuildAppWindow.buttonClickable;
-import static view.BuildAppWindow.buttonUnclickable;
-
-/**
- * The LoginView for when the user has just started the program, and not logged in.
- */
-public class LogoutView extends JPanel{
-
-    public LogoutView() {
-      // remove all buttons
-    }
-
-}
