+++ src/main/java/view/SearchView.java
@@ -1,9 +1,10 @@
 package view;
 
-import javax.swing.*;
-import java.awt.*;
+import app.BuildAppWindow;
 
-import static view.BuildAppWindow.*;
+import javax.swing.*;
+
+import static app.BuildAppWindow.*;
 
 /**
  * The LoginView for when the user has just started the program, and not logged in.
@@ -11,12 +12,18 @@ import static view.BuildAppWindow.*;
 public class SearchView extends JPanel{
 
     public SearchView() {
+
+        buttonsPanelVisible(buttonsPanel);
+
         buttonUnclickable(BuildAppWindow.searchButton);
         buttonClickable(BuildAppWindow.homeButton);
         buttonClickable(BuildAppWindow.calendarButton);
         buttonClickable(BuildAppWindow.bookmarksButton);
         buttonClickable(BuildAppWindow.settingsButton);
         buttonClickable(BuildAppWindow.logoutButton);
+
+        JLabel jlabel = new JLabel("Search for recipes");
+        this.add(jlabel);
     }
 
 }
